donkey_ears/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
cytoolz==0.12.0
Django==4.1.1
djangorestframework==3.14.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.17.0
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
Naked==0.1.31
netaddr==0.8.0
parsimonious==0.8.1
Pillow==9.2.0
protobuf==3.19.5
pycryptodome==3.15.0
pyrsistent==0.19.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.2.1
pywin32==305
PyYAML==6.0
requests==2.28.1
rlp==2.0.1
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
tzdata==2022.4
urllib3==1.26.12
varint==1.0.2
web3==5.31.1
websockets==9.1
yarl==1.8.1
